+++ Sources/MapboxNavigationCore/Map/Other/UIFont.swift
@@ -3,6 +3,10 @@ import UIKit
 @_spi(MapboxInternal)
 extension UIFont {
     public class var defaultRouteAnnotationTextFont: UIFont {
-        .systemFont(ofSize: 18)
+        .systemFont(ofSize: 18, weight: .semibold)
+    }
+
+    public class var defaultRouteAnnotationCaptionTextFont: UIFont {
+        .systemFont(ofSize: 16, weight: .regular)
     }
 }